R3 AR-1: extract parse/rewrite.rs from the parse coordinator
Completes the AR-1 god-module split. Prior phases already carved detection
(detect.rs), SHOW-clause parsing (show_clauses.rs), CREATE-body parsing
(create_body.rs), native-SQL emission (native_sql.rs), the FFI surface
(ffi.rs), cardinality inference (graph/cardinality.rs), and catalog DML
(catalog::writes) out of the old monolithic parse.rs. The remaining
rewrite-planning core still lived inline in parse/mod.rs.

This moves that core — RewriteAction / CreateMode, plan_rewrite, plan_ddl,
rewrite_alter / validate_alter, the name-extraction and quoted-string
helpers, function_name — plus its ~2900-line test module into a new
parse/rewrite.rs. parse/mod.rs (3717 -> 79 lines) is now a thin coordinator:
it declares the submodules, keeps the shared DdlKind enum and PARSE_* consts,
and re-exports the public API (plan_rewrite, extract_ddl_name, RewriteAction,
CreateMode) so every `crate::parse::*` and `super::` call site resolves
unchanged.

Pure move, no behavior change. Cross-module callers that reached
now-relocated private items get pub(crate): CreateMode::{from_kind,
or_replace, if_not_exists} (called by create_body / native_sql) and
extract_quoted_string (used by show_clauses). Also gates ffi.rs's
plan_rewrite import to the non-extension test config where it is actually
used (a latent unused-import in the extension build).
